pyproject.toml
[tool.poetry]
name = "sqlalchemy-test"
version = "0.1.0"
description = ""
authors = ["Renat Akzholov <akzholov.renat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.41"
asyncpg = "^0.26.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.4"
python-dotenv = "^0.21.0"
greenlet = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.982"
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
